Move simapp benchmarks to separate file
